Request the maximum possible session timeout when creating a session.
The OPC UA server always revises the session timeout, but previously adjusted it to its configured minimum timeout.
Now the timeout is adjusted to the server's maximum session timeout.
As long as no logic is in place that takes revised_session_timeout into consideration and sends periodic keepalives, this is a simple fix to combat many cases of silently dropped sessions.