Prevent leaking session activity timer threads (#87)
I noticed that our application was leaking threads (and with that file handles) so I did some investigations to understand what caused the problem.

In the end I pinned it down to an issue with the `session_state` being reused when using the reconnect logic.

The main logic in `spawn_session_activity_task` is set to execute on each interval until the `connection_state` is `ConnectionState::Finished(_)`: https://github.com/locka99/opcua/blob/master/client/src/session.rs#L1052-L1059

On itself this is perfectly fine, but the problem is that due to the interval in most cases the `connection_state` is already updated again by the new connection attemp before it was checked. So then when the interval triggers again, the `connection_state` is no longer `ConnectionState::Finished(_)` causing task to continue instead of abort.

By resetting the `session_state` within the `session`, a new `connection_state` is used for the new attempt and all the old reasources are dropped/closed. Tested this quite a bit and it seems to work perfect in our application. No more thread- or file handle leaks after this patch.