Use the current thread runtime for Tokio tasks (#81)
The default runtime is a threaded runtime which spawns a background
thread running a reactor instance and then creates a threadpool (with
the number of threads equal the the number of available CPU cores).

Now this creates a lot of threads, especially when multiple runtimes
are started, like in the client package. And while having a lot of
threads should not be a problem, Tokio needs to be able to communicate
between these threads so it (on Linux) creates unix pipes for that.
On an 16 core machine this resulted in around 144 handles per client.

On our system and the way we use multiple clients in our application
this caused us to hit the 1024 default open file handle limit. In most
cases that limit can easily be updated, but in our case that isn't an
option/solution. So I went looking at the code to see if there was an
"easy" fix and it seems there is one... As all the runtimes that are
created are actually created inside new dedicated threads, it seems
that using that single thread for the new runtime is a more efficient
and solution which seems to be inline with how things were intended to
work.

I made the required changes in this PR and hope this is indeed inline
with how you inteded to use Tokio. I excluded the runtime in the main
`run_server` function as I can imagine in that particular case it could
make sense to use a threaded runtime (now sure, can of cause make the
change for this one as well if you want).