Allow the client to use the server time
In some situations, were the server setup is managed by a 3rd party, it can be needed to “follow” the server time in the client in order to be able to setup a connection.

This solution is pretty much inline with what is documented [here](https://reference.opcfoundation.org/v104/Core/docs/Part6/6.3/), except in this solution only the time used in the client is adjusted and the host clock is left untouched.

And just to be clear, the offset is only used for the communication and security related parts between the client and the server. For the rest the actuall time will still be used.

I tested this against several 3rd party server and the change resulted in stable connection without any errors, reconnects or constantly renewing security channels.

Fixes #89