crypto/user_identity: Fix selection of security policy
In make_user_name_identity_token() the security policy to use when
authenticating with username and password is determined from the
user_token_policy and token_security_policy.

In our tests, if the server provides eg. Basic128Rsa15 as the
user token policy, the logic would select None as the policy and
subsequently not encrypt the password and fail to connect.

Inverting some of the logic appears to work (at least for our tests).