types/variant: fix cast to i32 that might result in negative number
When array_length is decoded and greater than zero, the configured
decoding options are checked, casting max_array_length to i32. The type
is a usize. We set max_array_length = usize::MAX, resulting the cast to
produce -1 (as an i32) and returning BadEncodingLimitsExceeded.

Given that array_length is cast to a usize immediately after this check,
it makes sense to first convert to it to usize and perform the
comparison on usize instead of i32.