Fix memory leaks in connection management code
This revamps the connection handling on both the client and the server
to guarantee that we never have rogue tokio tasks running.

This simplifies the connection management code by using native rust and tokio
future cancelling to make sure all concurrent code stops when a connection dies,
whatever the reason.

This allows removing a lot of complex connection state management code.

Fixes #199
Fixes #179