client/session: Allow setting requested_max_references_per_node
This indicates to servers how many nodes can be included in a response
to a browse request. Some servers abort a connection if the number of
nodes it's trying to inlude in a browse response is too many.

Change the signature of browse() to allow passing in a custom maximum
which can help as a workaround.