feat!: Drop the sample `web-client` temporarily
as it's currently not working for at least two different reasons:

Commit(43afb6a): 'Implement JSON serialization on most built-in data types.',

which broke the JS parsing and updating of subscribed values.
And the other being

Commit(eab4caa): 'Rewrite the client to be async all the way through',

which indirectly halted subscriptions for servers implemented with this crate (#324).

Re-adding this sample would be highly appreciated.
If possible with tests, which might prevent its recurring decay.