Make the ignore clock skew parameter follow the generic pattern
All other parameters take a value, while this one did not. Especially
when we (maybe) want to set the default to `true` one day, this would
become an issue (next to it not being consistent now).

Normally such a change would not be worth causing a backwards
incompatibility issue, but since the rewritten client isn't released yet
it feels like this is the best possible time to correct this.